Move withTimeout and friends to AsyncOperations
